Fixed various errors:
Fixed bug with short inline function in a function call through a variable when variable is not a function or has less arguments than position in which short inline function is written.
Fixed bug when generic member function pointer is taken.
Fixed error when member function pointer is taken when there are multiple member function overloads and the l-value is not a reference.
Improved code coverage by tests.