Fixed register conflict when separated phi webs of the same color exist in different sub-trees
When walking the dominance tree to allocate VM registers for live-in and live-out variables, a phi web might have received a VM register for its color.
Then, when stepping out of a sub-tree with the phi web and going into another, the same VM register can be assigned to a different virtual register because the phi web with the same color might be further ahead and we don't know about it yet.
So now, when phi web live range ends (while walking the block tree in dominance order), the mapping between color and register is reset.
Since the phi web was separated into standalone parts it is ok for different parts to have different VM registers.
If we were to reset the phi web coloring and assign it again these separated webs would have received different color anyway.